+++ profiler/eflect/data/MachineSample.java
@@ -11,15 +11,15 @@ public final class MachineSample implements Sample {
   private static final String STAT_FILE = String.join(File.separator, "/proc", "stat");
   private static final int[] JIFFY_INDICES = new int[] {1, 2, 3, 6, 7, 8, 9, 10};
 
-  private static int[] readMachineJiffies() {
-    int[] jiffies = new int[CPU_COUNT];
+  private static long[] readMachineJiffies() {
+    long[] jiffies = new long[CPU_COUNT];
     try (BufferedReader reader = new BufferedReader(new FileReader(new File(STAT_FILE)))) {
       reader.readLine();
       for (int i = 0; i < CPU_COUNT; i++) {
         String[] stats = reader.readLine().split(" ");
         int cpu = Integer.parseInt(stats[0].substring(3));
         for (int j: JIFFY_INDICES) {
-          jiffies[cpu] += Integer.parseInt(stats[j]);
+          jiffies[cpu] += Long.parseLong(stats[j]);
         }
       }
     } catch (Exception e) {
@@ -29,7 +29,7 @@ public final class MachineSample implements Sample {
     }
   }
 
-  private final int[] jiffies;
+  private final long[] jiffies;
   private final Instant timestamp;
 
   public MachineSample() {
@@ -37,7 +37,7 @@ public final class MachineSample implements Sample {
     this.timestamp = Instant.now();
   }
 
-  public int[] getJiffies() {
+  public long[] getJiffies() {
     return jiffies;
   }
